Initial commit
Hello

this is a small note for some of the changes I've added some debug information for rooms

White blocks indicate spikes
Black blocks indicate holes
Red boundary boxes is the boundaries for the actual dungeon room

I made this to speed up development with room creation, as otherwise it may as well be a guessing game :/
should you want to improve it just message me and I'll see what I can do, the code is located in - = TrapLocations.cs = - in the function OnGUI()

I've changed the totems to shoot with larger gaps of time between shots

We now have - = 6 = - rooms to work with [balancing not entirely assured]

I've re-added the better audio manager and have updated the music slightly

Enemy slimes have been added, but they lack any detection for blocks they cannot tresspass, so they will often keep running into walls or holes without stopping.

I do plan on adding a turret enemy that shoots in the 4 cardinal directions at once [it's past 4AM as of writing this so I think I have a good excuse]

I've yet to add the detection to what doors are and are not available, but that could easily be stored in a 4bit bitmask.

I trust you will be able to understand my code and blunders to help with future development!

I may make more rooms tomorrow!